orden de los archivos
Se acomodan los archivos para la sencilla descarga y ejecución de los programas
