--- caperucitaEDITADA.txt
@@ -1,1 +1,0 @@
-cuento de caperucita habia una vez una adorable nina que era querida por todo aquel que la conociera pero sobre todo por su abuelita y no quedaba nada que no le hubiera dado a la nina una vez le regalo una pequena caperuza o gorrito de un color rojo que le quedaba tan bien que ella nunca queria usar otra cosa asi que la empezaron a llamar caperucita roja un dia su madre le dijo ven caperucita roja aqui tengo un pastel y una botella de vino llevaselas en esta canasta a tu abuelita que esta enfermita y debil y esto le ayudara vete ahora temprano antes de que caliente el dia y en el camino camina tranquila y con cuidado no te apartes de la ruta no vayas a caerte y se quiebre la botella y no quede nada para tu abuelita y cuando entres a su dormitorio no olvides decirle buenos dias ah y no andes curioseando por todo el aposento no te preocupes hare bien todo dijo caperucita roja y tomo las cosas y se despidio carinosamente la abuelita vivia en el bosque como a un kilometro de su casa y no mas habia entrado caperucita roja en el bosque siempre dentro del sendero cuando se encontro con un lobo caperucita roja no sabia que esa criatura pudiera hacer algun dano y no tuvo ningun temor hacia el buenos dias caperucita roja dijo el lobo buenos dias amable lobo adonde vas tan temprano caperucita roja a casa de mi abuelita y que llevas en esa canasta pastel y vino ayer fue dia de hornear asi que mi pobre abuelita enferma va a tener algo bueno para fortalecerse y adonde vive tu abuelita caperucita roja como a medio kilometro mas adentro en el bosque su casa esta bajo tres grandes robles al lado de unos avellanos seguramente ya los habras visto contesto inocentemente caperucita roja el lobo se dijo en silencio a si mismo que criatura tan tierna que buen bocadito y sera mas sabroso que esa viejita asi que debo actuar con delicadeza para obtener a ambas facilmente entonces acompano a caperucita roja un pequeno tramo del camino y luego le dijo mira caperucita roja que lindas flores se ven por alla por que no vas y recoges algunas y yo creo tambien que no te has dado cuenta de lo dulce que cantan los pajaritos es que vas tan apurada en el camino como si fueras para la escuela mientras que todo el bosque esta lleno de maravillas caperucita roja levanto sus ojos y cuando vio los rayos del sol danzando aqui y alla entre los arboles y vio las bellas flores y el canto de los pajaros penso supongo que podria llevarle unas de estas flores frescas a mi abuelita y que le encantaran ademas aun es muy temprano y no habra problema si me atraso un poquito siempre llegare a buena hora y asi ella se salio del camino y se fue a cortar flores y cuando cortaba una veia otra mas bonita y otra y otra y sin darse cuenta se fue adentrando en el bosque mientras tanto el lobo aprovecho el tiempo y corrio directo a la casa de la abuelita y toco a la puerta quien es pregunto la abuelita caperucita roja contesto el lobo traigo pastel y vino abreme por favor mueve la cerradura y abre tu grito la abuelita estoy muy debil y no me puedo levantar el lobo movio la cerradura abrio la puerta y sin decir una palabra mas se fue directo a la cama de la abuelita y de un bocado se la trago y enseguida se puso ropa de ella se coloco un gorro se metio en la cama y cerro las cortinas mientras tanto caperucita roja se habia quedado colectando flores y cuando vio que tenia tantas que ya no podia llevar mas se acordo de su abuelita y se puso en camino hacia ella cuando llego se sorprendio al encontrar la puerta abierta y al entrar a la casa sintio tan extrano presentimiento que se dijo para si misma oh dios que incomoda me siento hoy y otras veces que me ha gustado tanto estar con abuelita entonces grito buenos dias pero no hubo respuesta asi que fue al dormitorio y abrio las cortinas alli parecia estar la abuelita con su gorro cubriendole toda la cara y con una apariencia muy extrana oh abuelita dijo que orejas tan grandes que tienes es para oirte mejor mi nina fue la respuesta pero abuelita que ojos tan grandes que tienes son para verte mejor querida pero abuelita que brazos tan grandes que tienes para abrazarte mejor y que boca tan grande que tienes para comerte mejor y no habia terminado de decir lo anterior cuando de un salto salio de la cama y se trago tambien a caperucita roja y colorin colorado este cuento se ha acabado
