+++ proyectoF/caperucita.txt
@@ -1,0 +1,13 @@
+Cuento de caperucita
+
+Había una vez una adorable niña que era querida por todo aquél que la conociera, pero sobre todo por su abuelita, y no quedaba nada que no le hubiera dado a la niña. Una vez le regaló una pequeña caperuza o gorrito de un color rojo, que le quedaba tan bien que ella nunca quería usar otra cosa, así que la empezaron a llamar Caperucita Roja. Un día su madre le dijo: "Ven, Caperucita Roja, aquí tengo un pastel y una botella de vino, llévaselas en esta canasta a tu abuelita que esta enfermita y débil y esto le ayudará. Vete ahora temprano, antes de que caliente el día, y en el camino, camina tranquila y con cuidado, no te apartes de la ruta, no vayas a caerte y se quiebre la botella y no quede nada para tu abuelita. Y cuando entres a su dormitorio no olvides decirle, "Buenos días," ah, y no andes curioseando por todo el aposento."
+
+"No te preocupes, haré bien todo," dijo Caperucita Roja, y tomó las cosas y se despidió cariñosamente. La abuelita vivía en el bosque, como a un kilómetro de su casa. Y no más había entrado Caperucita Roja en el bosque, siempre dentro del sendero, cuando se encontró con un lobo. Caperucita Roja no sabía que esa criatura pudiera hacer algún daño, y no tuvo ningún temor hacia él. "Buenos días, Caperucita Roja," dijo el lobo. "Buenos días, amable lobo." - "¿Adonde vas tan temprano, Caperucita Roja?" - "A casa de mi abuelita." - "¿Y qué llevas en esa canasta?" - "Pastel y vino. Ayer fue día de hornear, así que mi pobre abuelita enferma va a tener algo bueno para fortalecerse." - "¿Y adonde vive tu abuelita, Caperucita Roja?" - "Como a medio kilómetro más adentro en el bosque. Su casa está bajo tres grandes robles, al lado de unos avellanos. Seguramente ya los habrás visto," contestó inocentemente Caperucita Roja.
+
+El lobo se dijo en silencio a sí mismo: "¡Qué criatura tan tierna! qué buen bocadito - y será más sabroso que esa viejita. Así que debo actuar con delicadeza para obtener a ambas fácilmente." Entonces acompañó a Caperucita Roja un pequeño tramo del camino y luego le dijo: "Mira Caperucita Roja, que lindas flores se ven por allá, ¿por qué no vas y recoges algunas? Y yo creo también que no te has dado cuenta de lo dulce que cantan los pajaritos. Es que vas tan apurada en el camino como si fueras para la escuela, mientras que todo el bosque está lleno de maravillas."
+
+Caperucita Roja levantó sus ojos, y cuando vio los rayos del sol danzando aquí y allá entre los árboles, y vio las bellas flores y el canto de los pájaros, pensó: "Supongo que podría llevarle unas de estas flores frescas a mi abuelita y que le encantarán. Además, aún es muy temprano y no habrá problema si me atraso un poquito, siempre llegaré a buena hora." Y así, ella se salió del camino y se fue a cortar flores. Y cuando cortaba una, veía otra más bonita, y otra y otra, y sin darse cuenta se fue adentrando en el bosque. Mientras tanto el lobo aprovechó el tiempo y corrió directo a la casa de la abuelita y tocó a la puerta. "¿Quién es?" preguntó la abuelita. "Caperucita Roja," contestó el lobo. "Traigo pastel y vino. Ábreme, por favor." - "Mueve la cerradura y abre tú," gritó la abuelita, "estoy muy débil y no me puedo levantar." El lobo movió la cerradura, abrió la puerta, y sin decir una palabra más, se fue directo a la cama de la abuelita y de un bocado se la tragó. Y enseguida se puso ropa de ella, se colocó un gorro, se metió en la cama y cerró las cortinas.
+
+Mientras tanto, Caperucita Roja se había quedado colectando flores, y cuando vio que tenía tantas que ya no podía llevar más, se acordó de su abuelita y se puso en camino hacia ella. Cuando llegó, se sorprendió al encontrar la puerta abierta, y al entrar a la casa, sintió tan extraño presentimiento que se dijo para sí misma: "¡Oh Dios! que incómoda me siento hoy, y otras veces que me ha gustado tanto estar con abuelita." Entonces gritó: "¡Buenos días!," pero no hubo respuesta, así que fue al dormitorio y abrió las cortinas. Allí parecía estar la abuelita con su gorro cubriéndole toda la cara, y con una apariencia muy extraña. "¡!Oh, abuelita!" dijo, "qué orejas tan grandes que tienes." - "Es para oírte mejor, mi niña," fue la respuesta. "Pero abuelita, qué ojos tan grandes que tienes." - "Son para verte mejor, querida." - "Pero abuelita, qué brazos tan grandes que tienes." - "Para abrazarte mejor." - "Y qué boca tan grande que tienes." - "Para comerte mejor." Y no había terminado de decir lo anterior, cuando de un salto salió de la cama y se tragó también a Caperucita Roja.
+
+Y colorín colorado, este cuento se ha acabado.
